Add style store
