+++ web/src/components/sidebars/LayerStyle.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
-import { rasterColormaps, updateLayerStyles } from '@/layerStyles';
-import { useLayerStore } from '@/store/layer';
+import { useStyleStore } from '@/store/style';
 import { Layer } from '@/types';
 import { computed, watch } from 'vue';
 import _ from 'lodash';
@@ -9,23 +8,23 @@ const props = defineProps<{
   layer: Layer;
 }>();
 
-const layerStore = useLayerStore();
+const styleStore = useStyleStore();
 
 const styleKey = computed(() => {
     return `${props.layer.id}.${props.layer.copy_id}`;
 })
 
 const currentStyle = computed(() => {
-    if (!layerStore.selectedLayerStyles[styleKey.value]) {
-        layerStore.selectedLayerStyles[styleKey.value] = {
+    if (!styleStore.selectedLayerStyles[styleKey.value]) {
+        styleStore.selectedLayerStyles[styleKey.value] = {
             visible: true,
             opacity: 1,
         }
     }
-    if (!layerStore.selectedLayerStyles[styleKey.value].colormap_range) {
-        layerStore.selectedLayerStyles[styleKey.value].colormap_range = dataRange.value
+    if (!styleStore.selectedLayerStyles[styleKey.value].colormap_range) {
+        styleStore.selectedLayerStyles[styleKey.value].colormap_range = dataRange.value
     }
-    return layerStore.selectedLayerStyles[styleKey.value];
+    return styleStore.selectedLayerStyles[styleKey.value];
 });
 
 const showRasterOptions = computed(() => {
@@ -55,7 +54,7 @@ function getFrameInputWidth(value: number) {
 }
 
 watch(currentStyle, _.debounce(() => {
-    updateLayerStyles(props.layer)
+    styleStore.updateLayerStyles(props.layer)
 }, 100), {deep: true})
 </script>
 
@@ -104,7 +103,7 @@ watch(currentStyle, _.debounce(() => {
                 </v-label>
                 <v-select
                     v-model="currentStyle.colormap"
-                    :items="rasterColormaps"
+                    :items="styleStore.rasterColormaps"
                     label="Color Map"
                     density="compact"
                 />
